fix: ContainerSetSlot slot is i16 on the wire, not varint
M6 took javap's field type ('int slot') at face value and shipped
slot as a varint. Vanilla 26.1.2 encodes it as a short (same as
the symmetric serverbound SetCarriedItem.slot we already have
right), so every post-edit ContainerSetSlot slid the downstream
ItemStack payload — vanilla clients dropped the connection with
'Failed to decode packet clientbound/minecraft:container_set_slot'
the moment the M6 inventory decrement fired.

Caught on the M9 manual gate. Pure M6 carry-over, no M9 logic
involved.

+++ crates/mc-net/src/play.rs
@@ -977,7 +977,7 @@ where
         &ClientboundContainerSetSlot {
             container_id: 0,
             state_id: state.inventory_state_id,
-            slot: (PlayerInventory::HOTBAR_BASE + held_slot as usize) as i32,
+            slot: (PlayerInventory::HOTBAR_BASE + held_slot as usize) as i16,
             item_stack: new_slot_value,
         },
         Compression::Disabled,
